rskrza1: Allow 32MB address of SPI Flash for QSPI XIP
If left at the POR default value, only 16MB of physical SPI flash can be
accessed because bits 25-31 of the SPI Flash addresses are fixed to 0.
Best we can do is set DREAR:EAC to 1 which will give use 25 bits of
physical address space accessibility.
This allows for a virtual address range of 32MB for single flash
or 64MB for dual flash.