rz_spi: Properly end SPI session
Doing the dummy XIP reads was a hack to get things working right when
transitioning from SPI mode to XIP mode. But, it turns out that the issue
was chip select line was left low by the upper layer because it assumed
SPI_XFER_END would release it.
So...now we explicitly do that.